requirements.txt
biopython>=1.57
click~=8.1.8
loguru~=0.7.3
pytest~=8.3.5